Fix item_teaser templatags now use TBlauncher to get the data
