+++ transport_nantes/asso_tn/templates/asso_tn/index.html
@@ -3,7 +3,6 @@
 {% load mpanels %}
 {% load newsletter %}
 {% load launcher %}
-{% load item_teaser %}
 {% block content %}
 {% item_teaser 'ligne-johanna-rolland-pour-plus-de-mobilite' %}
 <div class="background-ad-hoc-blue">
